Implemented validation for the Id field

+++ src/main/java/edu/fontysmaua/tournamentapi/business/impl/DeleteTournamentUseCaseImpl.java
@@ -11,6 +11,12 @@ public class DeleteTournamentUseCaseImpl implements DeleteTournamentUseCase {
     private final TournamentRepository tournamentRepository;
 
     public Long delete(Long id) {
+        if (id == null) {
+            throw new IllegalArgumentException("ID cannot be null.");
+        }
+        if(id <= 0){
+            throw new IllegalArgumentException("ID must be greater than 0.");
+        }
         tournamentRepository.deleteById(id);
         return id;
     }
